Scripts are agnostic of Github pages
Github pages are just one way of deploying the app so changing the
local folder name in which the app is built, and adapting the
scripts so that only deploy to Github pages has pages in its name.
Deploy also no longer require a setup phase to be run once, and
pushed history on gh-pages branch is always crushed.